Prefer server ciphers for maximum SSL security.
When negotiating SSL, the cipher to use is chosen from the list of
useable ciphers, that is, the ones listed in the ssl_ciphers setting
that are also supported by the client.

With this setting turned on, the chosen cipher is the first useable
cipher in the server's list, not the client's. This lets us choose
which ciphers to prefer with the ordering of the ssl_ciphers setting
(which is already looking good).

This is further to the perfect forward secrecy tweaks in 324f85c.