install s3cmd on all systems
The version available in ubuntu 14.04 requires AWS keys to be written to
a config file, which is a giant pain.

This newer version allows them to be passed as cli options. That's less
secure (they may be visible in process lists) but we control the server
so I can handle that. It'll also make it easier to shell out from ruby
with a complete config