Handle puma socket in nginx dep
We has a special case in the nginx dep for the analytics app, which up until recently was the only app we had running on puma. This won't be the case for much longer.

This PR updates the nginx dep to automatically detect puma apps and use the correct socket name.