stop hard coding the number of workers and threads for unicorn and puma
We'd like the option to control these values in configuration that
lives inside each app repo.

In pumas case, we normally do a phased restart (USR1) during a deploy.
It's worth noting that this style of restart *doesn't* pick up changes
in the configuration file, so deploys with change to the puma config
will need to do a regular restart (USR2)