Add a script that logs key puma metrics to a local statsd server
Traditionally we've deployed our apps with unicorn, and the key thing to
monitor when watching for capacity issues is the socket between nginx
and unicorn. If that backs up, users experience request queuing and
we're at risk of serving 502s (fail whale).

For puma, we already have monitoring on the socket between nginx and
puma - but that's less interesting. Puma doesn't queue requests on that
socket, so it almost always looks empty.

puma has a master process and 1 or more worker processes, and requests
appear to be queued inside the master process. So, to monitor the number
of active and queued requests we need to ask the master process for some
stats.

This script will do so every second, and send the key stats to a local
statsd server. From their, our monitoring pipeline will send the data to
librato where we can graph it on a dashboard.

I've incubated this script in the analytics codebase [1], so I'm pretty
confident it works well. This change just bakes it into our provisioning
so it's automatically used for all our puma apps.

Props to Tim Lucas and Sam Cochran for the basic structure:

    https://gist.github.com/toolmantim/0e76d0377bb23974b06f
    https://github.com/sj26/puma-plugin-systemd

[1] https://github.com/conversation/tc-analytics/pull/576